Refactoring file downloading utilities
Consolidate http connection utility methods into
org.omegat.util.net.HttpConnectionUtils class.
