+++ src/org/omegat/gui/dialogs/DictionaryInstallerDialog.java
@@ -44,7 +44,7 @@ import javax.swing.JOptionPane;
 import javax.swing.SwingWorker;
 
 import org.omegat.core.spellchecker.DictionaryManager;
-import org.omegat.core.spellchecker.DictionaryManager.FlakyDownloadException;
+import org.omegat.util.net.HttpConnectionUtils.FlakyDownloadException;
 import org.omegat.util.Log;
 import org.omegat.util.OStrings;
 import org.omegat.util.gui.StaticUIUtils;
